Isolating Opening Hours component
